+++ DataDefinitions/Body.cs
@@ -104,6 +104,23 @@ namespace EddiDataDefinitions
         }
         [JsonIgnore] private bool _mappedEfficiently;
 
+        /// <summary>Whether we're the first commander to step foot on this body</summary>
+        [PublicAPI("Whether a commander has already claimed the first footfall for this body")]
+        public bool? alreadyfootfalled { get; set; }
+
+        [PublicAPI( "When you claimed the first footfall for this body, if you have" ), JsonIgnore]
+        public long? footfalled => _footfalledDateTime is null
+            ? null
+            : (long?)Dates.fromDateTimeToSeconds( (DateTime)_footfalledDateTime );
+
+        [JsonProperty( nameof( footfalled ) )]
+        public DateTime? footfalledDateTime
+        {
+            get => _footfalledDateTime;
+            set { _footfalledDateTime = value; OnPropertyChanged(); }
+        }
+        [JsonIgnore] private DateTime? _footfalledDateTime;
+
         /// <summary>The estimated value of the body</summary>
         [PublicAPI, JsonIgnore]
         public long estimatedvalue => scannedDateTime == null 
@@ -434,7 +451,7 @@ namespace EddiDataDefinitions
         public ReserveLevel reserveLevel { get; set; } = ReserveLevel.None;
 
         /// <summary> Planet or Moon definition </summary>
-        public Body(string bodyName, long? bodyId, string systemName, ulong systemAddress, List<IDictionary<string, object>> parents, decimal? distanceLs, bool? tidallylocked, TerraformState terraformstate, PlanetClass planetClass, AtmosphereClass atmosphereClass, List<AtmosphereComposition> atmosphereCompositions, Volcanism volcanism, decimal? earthmass, decimal? radiusKm, decimal gravity, decimal? temperatureKelvin, decimal? pressureAtm, bool? landable, List<MaterialPresence> materials, List<SolidComposition> solidCompositions, decimal? semimajoraxisLs, decimal? eccentricity, decimal? orbitalinclinationDegrees, decimal? periapsisDegrees, decimal? orbitalPeriodDays, decimal? rotationPeriodDays, decimal? axialtiltDegrees, List<Ring> rings, ReserveLevel reserveLevel, bool? alreadydiscovered, bool? alreadymapped)
+        public Body(string bodyName, long? bodyId, string systemName, ulong systemAddress, List<IDictionary<string, object>> parents, decimal? distanceLs, bool? tidallylocked, TerraformState terraformstate, PlanetClass planetClass, AtmosphereClass atmosphereClass, List<AtmosphereComposition> atmosphereCompositions, Volcanism volcanism, decimal? earthmass, decimal? radiusKm, decimal gravity, decimal? temperatureKelvin, decimal? pressureAtm, bool? landable, List<MaterialPresence> materials, List<SolidComposition> solidCompositions, decimal? semimajoraxisLs, decimal? eccentricity, decimal? orbitalinclinationDegrees, decimal? periapsisDegrees, decimal? orbitalPeriodDays, decimal? rotationPeriodDays, decimal? axialtiltDegrees, List<Ring> rings, ReserveLevel reserveLevel, bool? alreadydiscovered, bool? alreadymapped, bool? alreadyfootfalled)
         {
             this.bodyname = bodyName;
             this.bodyType = (bool)parents?.Exists(p => p.ContainsKey("Planet"))
@@ -478,6 +495,7 @@ namespace EddiDataDefinitions
             // Scan details
             this.alreadydiscovered = alreadydiscovered;
             this.alreadymapped = alreadymapped;
+            this.alreadyfootfalled = alreadyfootfalled;
         }
 
         // Additional calculated planet and moon statistics