+++ Events/DisembarkEvent.cs
@@ -9,8 +9,11 @@ namespace EddiEvents
     {
         public const string NAME = "Disembark";
         public const string DESCRIPTION = "Triggered when you transition from a ship or SRV to on foot";
-        public const string SAMPLE = "{ \"timestamp\":\"2021-05-03T21:47:38Z\", \"event\":\"Disembark\", \"SRV\":false, \"Taxi\":false, \"Multicrew\":false, \"ID\":6, \"StarSystem\":\"Firenses\", \"SystemAddress\":2868635379121, \"Body\":\"Roberts Gateway\", \"BodyID\":44, \"OnStation\":true, \"OnPlanet\":false, \"StationName\":\"Roberts Gateway\", \"StationType\":\"Coriolis\", \"MarketID\":3221636096 }";
-
+        public static readonly string[] SAMPLES =
+        {
+            "{ \"timestamp\":\"2021-05-03T21:47:38Z\", \"event\":\"Disembark\", \"SRV\":false, \"Taxi\":false, \"Multicrew\":false, \"ID\":6, \"StarSystem\":\"Firenses\", \"SystemAddress\":2868635379121, \"Body\":\"Roberts Gateway\", \"BodyID\":44, \"OnStation\":true, \"OnPlanet\":false, \"StationName\":\"Roberts Gateway\", \"StationType\":\"Coriolis\", \"MarketID\":3221636096 }"
+        };
+            
         [PublicAPI("The name of the star system where the commander is disembarking")]
         public string systemname { get; }
 
@@ -49,6 +52,9 @@ namespace EddiEvents
 
         [PublicAPI("True if disembarking to a planet")]
         public bool? onplanet { get; }
+        
+        [PublicAPI( "True if disembarking on a planet with no previously registered first footfall" )]
+        public bool firstfootfall { get; set; }
 
         // Not intended to be user facing
         public int? fromLocalId { get; }